CU-663: Create ISO6709LocationMapper and refactor location to double from float

+++ domain/src/main/kotlin/mega/privacy/android/domain/entity/camerauploads/CameraUploadsRecord.kt
@@ -34,8 +34,8 @@ data class CameraUploadsRecord(
     val originalFingerprint: String,
     val generatedFingerprint: String?,
     val tempFilePath: String,
-    val latitude: Float? = null,
-    val longitude: Float? = null,
+    val latitude: Double? = null,
+    val longitude: Double? = null,
     val existsInTargetNode: Boolean? = null,
     val existingNodeId: NodeId? = null,
 )
